refactor(account-switcher): separate mobile & web logic
